Added initial system-status script for Raspberry Pi 5
This script displays the current status of the Raspberry Pi system. This is just the initial script which will print the main temperatures, clocks, and voltages. There are also additional temperatures it can show if lm-sensors and nvme-cli are installed. 
To use this create a folder called 'bin' in your home folder and put this file into that folder. After initially creating the bin folder you need to reboot, but after that you can just type system-status into terminal to run the script.
Example path to where this file goes: (/home/user/bin/)

This technically can also go into the root bin folder at /bin, but it is not recommended.