Finished making README with a full description of the repo.
I added a detailed description of the 2 main scripts in this repository. This will describe what each script does along with what it outputs to terminal. There is also and example output at the bottom. I may update this later once I make the system-status script also print the RTC battery voltage if one is available.