Add standalone RTC battery check script
This is a simple standalone script which uses the RTC driver to check the RTC battery reading via the system files. It will check if there is a battery, and it will check if the trickle charging is enabled. The script will just print the battery voltage and the charging voltage if available. Currently uses the fixed sys file path for the RTC battery readings from my raspberry pi, so I'm not sure if this file path will be the same for other people. I will try updating this with a section that checks the file system for the correct path instead of hard coding it so that it works for everyone.