Add misc script that outputs power consumption from the PMIC ADC
This script just takes every available voltage/current pair that is measured by the PMIC and calculates the power consumption according to those values. This does not give the full power consumption of the system. This only takes into account the values found when running 'vcgencmd pmic_read_adc' which is mostly all just the logic level rails like the HDMI power, 1.8v, 0.8v, and 3.3v power. The output power will pretty much always be pretty small unless you are doing a ton of general IO operations. This is just a random script I made and it's not very useful for anything.