Added section for RTC battery status.
I added a section before the NVME section which is just a copy of the rtc-check script, but less it gives a less verbose output. This added section just checks if an RTC battery is connected, and if so it prints it's current voltage. It will then also check if the RTC trickle charging is enabled, and if so it will print the currently set trickle charge voltage. This will also be able to detect if a battery is connected but dead.