Updated README with newly added scripts and features.
I added an rtc-check script along with the check-pmic-power script. In addition to this I also hard coded a less verbose/wordy version of the rtc-check  script in the main system-status script to show the RTC battery voltage after the supply voltage. Because of this I updated the description and example output of the main system-status script. I also added descriptions and example outputs for the new standalone scripts I added.